Updated how collision checking maintains a list of frame pairs to test. Introduced rulebased ProximitySetup as replacement for CollisionSetup.
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@2893 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/trajectory/Trajectory.hpp
@@ -123,7 +123,7 @@ public:
     std::vector<T> getPath(double dt, bool uniform = true) {
         std::vector<T> path;
         if (uniform) {
-            int steps = std::ceil(duration()/dt);
+            int steps = (int)std::ceil(duration()/dt);
             double delta = duration()/steps;
             for (double t = 0; t<=duration(); t += delta) {
                 path.push_back(x(t));
